pyproject.toml
[project]
name = "mcp-step"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "envyaml>=1.10.211231",
    "httpx>=0.28.1",
    "ipython>=9.4.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-mcp-adapters>=0.1.8",
    "langgraph>=0.5.1",
    "mcp[cli]>=1.10.1",
    "oci>=2.154.3",
    "oracledb>=3.2.0",
    "python-box>=7.3.2",
    "python-dotenv>=1.1.1",
    "slack-bolt>=1.23.0",
]
